Create HW7
Demonstrates the use of triggers, tables, and packages for tracking changes and retrieving information from a database. It includes creating tables and triggers to log deletions and updates for Employees, Emp, and departments tables, and it also demonstrates a package (Pack_Dept) with functions to retrieve department information by department ID or name.